Add QrCodeViewModel to CameraX-MlKit.
