+++ CameraX-MLKit/app/src/main/java/com/example/camerax_mlkit/QrCodeDrawable.kt
@@ -10,11 +10,8 @@ import com.google.mlkit.vision.barcode.common.Barcode
 
 /**
  * A Drawable that handles displaying a QR Code's data and a bounding box around the QR code.
- *
- * As is, this class only handles displaying the QR Code data if it's a URL. Other data types
- * can be handled by adding more cases of Barcode.TYPE_URL in the init block.
  */
-class QrCodeDrawable(barcode: Barcode) : Drawable() {
+class QrCodeDrawable(qrCodeViewModel: QrCodeViewModel) : Drawable() {
     private val boundingRectPaint = Paint().apply {
         style = Paint.Style.STROKE
         color = Color.YELLOW
@@ -34,49 +31,24 @@ class QrCodeDrawable(barcode: Barcode) : Drawable() {
         textSize = 36F
     }
 
-    private var boundingRect: Rect = barcode.boundingBox!!
+    private val qrCodeViewModel = qrCodeViewModel
     private val contentPadding = 25
-
-    private var qrContent: String = ""
-    private var textWidth: Int = 0
-    var qrCodeTouchCallback = { v: View, e: MotionEvent -> false} //no-op
-
-    init {
-        when (barcode.valueType) {
-            Barcode.TYPE_URL -> {
-                qrContent = barcode.url!!.url!!
-                qrCodeTouchCallback = { v: View, e: MotionEvent ->
-                    if (e.action == MotionEvent.ACTION_DOWN && boundingRect.contains(e.getX().toInt(), e.getY().toInt())) {
-                        val openBrowserIntent = Intent(Intent.ACTION_VIEW)
-                        openBrowserIntent.data = Uri.parse(qrContent)
-                        v.context.startActivity(openBrowserIntent)
-                    }
-                    true // return true from the callback to signify the event was handled
-                }
-            }
-            // Add other QR Code types here to handle other types of data,
-            // like Wifi credentials.
-            else -> {
-                qrContent = "Unsupported data type: ${barcode.rawValue.toString()}"
-            }
-        }
-        textWidth = contentTextPaint.measureText(qrContent).toInt()
-    }
+    private var textWidth = contentTextPaint.measureText(qrCodeViewModel.qrContent).toInt()
 
     override fun draw(canvas: Canvas) {
-        canvas.drawRect(boundingRect, boundingRectPaint)
+        canvas.drawRect(qrCodeViewModel.boundingRect, boundingRectPaint)
         canvas.drawRect(
             Rect(
-                boundingRect.left,
-                boundingRect.bottom + contentPadding/2,
-                boundingRect.left + textWidth + contentPadding*2,
-                boundingRect.bottom + contentTextPaint.textSize.toInt() + contentPadding),
+                qrCodeViewModel.boundingRect.left,
+                qrCodeViewModel.boundingRect.bottom + contentPadding/2,
+                qrCodeViewModel.boundingRect.left + textWidth + contentPadding*2,
+                qrCodeViewModel.boundingRect.bottom + contentTextPaint.textSize.toInt() + contentPadding),
             contentRectPaint
         )
         canvas.drawText(
-            qrContent,
-            (boundingRect.left + contentPadding).toFloat(),
-            (boundingRect.bottom + contentPadding*2).toFloat(),
+            qrCodeViewModel.qrContent,
+            (qrCodeViewModel.boundingRect.left + contentPadding).toFloat(),
+            (qrCodeViewModel.boundingRect.bottom + contentPadding*2).toFloat(),
             contentTextPaint
         )
     }
